Added CLI integration tests
This change refactors how database tests are run. There are two
initialization methods that return a lock guard that controls the flow
of other database tests. Safe tests are written using a transaction to
isolate all of the databse calls, and are therefore safe to run
simultaneously as other safe tests. Unsafe tests are written in a way
that they try to clean themselves up, but the test environment will
automatically clear the database before each unsafe test.

The last change is that tests are now run in their own database, and
during the first database test's initialization, all migrations are
unwound and then performed to give the database a clean slate.

These tests mean that the database migrations themselves are now run as
part of the test harness, which is another added bonus.